[avswitch] workaround gb 50/60 issue

+++ lib/python/Components/AVSwitch.py
@@ -155,11 +155,11 @@ class AVSwitch:
 			mode_60 = mode_50
 
 		mode_etc = None
-		if os.path.exists('/proc/stb/video/videomode_50hz'):
+		if os.path.exists('/proc/stb/video/videomode_50hz') and not getBoxType().startswith('gb'):
 			f = open("/proc/stb/video/videomode_50hz", "w")
 			f.write(mode_50)
 			f.close()
-		if os.path.exists('/proc/stb/video/videomode_60hz'):
+		if os.path.exists('/proc/stb/video/videomode_60hz') and not getBoxType().startswith('gb'):
 			f = open("/proc/stb/video/videomode_60hz", "w")
 			f.write(mode_60)
 			f.close()
